refactor: Move Account to separate file
